1. Organized packages; 2. deleted rest/login authentication;  3. Updated README.md
