+++ src/integration-test/java/com/jaspersoft/jasperserver/jaxrs/client/apiadapters/core/JasperserverRestClientIT.java
@@ -1,6 +1,9 @@
 package com.jaspersoft.jasperserver.jaxrs.client.apiadapters.core;
 
 import com.jaspersoft.jasperserver.jaxrs.client.core.*;
+import com.jaspersoft.jasperserver.jaxrs.client.core.enums.AuthenticationType;
+import com.jaspersoft.jasperserver.jaxrs.client.core.enums.JRSVersion;
+import com.jaspersoft.jasperserver.jaxrs.client.core.enums.MimeType;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
@@ -28,14 +31,6 @@ public class JasperserverRestClientIT {
 
     @Test
     public void should_return_session_via_j_sucurity_check() {
-        config.setAuthenticationType(AuthenticationType.SPRING);
-        session = client.authenticate("superuser", "superuser");
-        assertNotNull(session);
-        assertNotNull(session.getStorage().getSessionId());
-    }
-
-    @Test
-    public void should_return_session_via_rest_login() {
         session = client.authenticate("superuser", "superuser");
         assertNotNull(session);
         assertNotNull(session.getStorage().getSessionId());
