+++ src/main/java/com/jaspersoft/jasperserver/jaxrs/client/core/Session.java
@@ -33,6 +33,7 @@ import com.jaspersoft.jasperserver.jaxrs.client.apiadapters.query.QueryExecutorS
 import com.jaspersoft.jasperserver.jaxrs.client.apiadapters.reporting.ReportingService;
 import com.jaspersoft.jasperserver.jaxrs.client.apiadapters.resources.ResourcesService;
 import com.jaspersoft.jasperserver.jaxrs.client.apiadapters.thumbnails.ThumbnailsService;
+import com.jaspersoft.jasperserver.jaxrs.client.core.enums.AuthenticationType;
 import com.jaspersoft.jasperserver.jaxrs.client.core.exceptions.handling.DefaultErrorHandler;
 import javax.ws.rs.client.WebTarget;
 import javax.ws.rs.core.Response;
