+++ src/main/java/com/jaspersoft/jasperserver/jaxrs/client/core/enums/ResourceMediaType.java
@@ -19,7 +19,7 @@
  * along with this program.&nbsp; If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.jaspersoft.jasperserver.jaxrs.client.core;
+package com.jaspersoft.jasperserver.jaxrs.client.core.enums;
 
 import java.util.Arrays;
 import java.util.List;
