+++ src/main/java/com/jaspersoft/jasperserver/jaxrs/client/core/enums/ResponseStatus.java
@@ -19,7 +19,7 @@
  * along with this program.&nbsp; If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.jaspersoft.jasperserver.jaxrs.client.core;
+package com.jaspersoft.jasperserver.jaxrs.client.core.enums;
 
 public final class ResponseStatus {
