+++ src/test/java/com/jaspersoft/jasperserver/jaxrs/client/core/SessionTest.java
@@ -15,6 +15,7 @@ import com.jaspersoft.jasperserver.jaxrs.client.apiadapters.reporting.ReportingS
 import com.jaspersoft.jasperserver.jaxrs.client.apiadapters.resources.ResourcesService;
 import com.jaspersoft.jasperserver.jaxrs.client.apiadapters.serverInfo.ServerInfoService;
 import com.jaspersoft.jasperserver.jaxrs.client.apiadapters.thumbnails.ThumbnailsService;
+import com.jaspersoft.jasperserver.jaxrs.client.core.enums.AuthenticationType;
 import com.jaspersoft.jasperserver.jaxrs.client.core.exceptions.RequestedRepresentationNotAvailableForResourceException;
 import junit.framework.Assert;
 import org.mockito.Mockito;
